+++ spec/integrations/grpc/perform_spec.rb
@@ -47,5 +47,48 @@ describe "client messages" do
         expect(subject.stop_streams).to eq false
       end
     end
+
+    describe ".state_attr_accessor" do
+      let(:data) { {action: "itick"} }
+
+      it "track attrs in the channel state" do
+        first_call = service.command(request)
+
+        expect(first_call).to be_success
+        expect(first_call.transmissions.size).to eq 1
+        expect(first_call.transmissions.first).to include({"result" => 1}.to_json)
+        expect(first_call.istate.to_h).not_to be_empty
+
+        first_state = first_call.istate
+
+        request.istate = first_state
+
+        second_call = service.command(request)
+
+        expect(second_call).to be_success
+        expect(second_call.transmissions.size).to eq 1
+        expect(second_call.transmissions.first).to include({"result" => 2}.to_json)
+        expect(second_call.istate.to_h).not_to be_empty
+        expect(second_call.istate).not_to eq(first_state)
+      end
+
+      context "with complex values" do
+        let!(:another_user) { User.create!(name: "alice") }
+
+        it "uses global id when possible and JSON otherwise" do
+          request.data = {action: "chat_with", user_id: another_user.id, topics: {oss: 1, ruby: 2}}.to_json
+          first_call = service.command(request)
+          expect(first_call).to be_success
+
+          request.istate = first_call.istate
+          request.data = {action: "send_message", text: "boom!", topic: :ruby}.to_json
+
+          second_call = service.command(request)
+          expect(second_call.transmissions.size).to eq 1
+          expect(second_call.transmissions.first).to include({"user" => "alice", "topic" => 2, "message" => "boom!"}.to_json)
+          expect(second_call.istate.to_h).to be_empty
+        end
+      end
+    end
   end
 end